feat : 자동차 객체 구현
