+++ src/main/java/racingcarwinner/Car.java
@@ -1,22 +1,26 @@
 package racingcarwinner;
 
-import racingcar.RandomGenerator;
-
 public class Car {
 
-    private final int ENABLE_MOVE_NUMBER = 4;
+
     private int travelDistance;
-    private RandomGenerator randomGenerator = new RandomGenerator();
+    private String name;
+
+    private Car(){ }
+
+    public Car(String name){
+        this.name = name;
+    }
 
     public void move(){
         this.travelDistance++;
     }
 
-    public boolean isOrAboveEnableNumber(int randomNumber){
-        return randomNumber >= ENABLE_MOVE_NUMBER;
-    }
-
     public int getTravelDistance(){
         return this.travelDistance;
+    }
+
+    public String getName(){
+        return this.name;
     }
 }
