+++ src/main/java/racingcarwinner/GameController.java
@@ -1,0 +1,2 @@
+package racingcarwinner;public class GameController {
+}
