+++ src/main/java/racingcarwinner/InputView.java
@@ -1,0 +1,2 @@
+package racingcarwinner;public class InputView {
+}
