+++ src/main/java/racingcarwinner/Winner.java
@@ -1,0 +1,2 @@
+package racingcarwinner;public class Winner {
+}
