+++ src/main/java/racingcarwinner/WinnerPicker.java
@@ -1,0 +1,2 @@
+package racingcarwinner;public class WinnerPicker {
+}
